+++ types/components/initClipboard.d.ts
@@ -8,7 +8,7 @@
  * <!--begin::Input group-->
  * <div class="input-group">
  *   <!--begin::Input-->
- *   <input id="target" type="text" class="form-control" placeholder="name@example.com" value="name@example.com" />
+ *   <input id="target" type="text" class="form-control" placeholder="name@example.com" value="name@example.com">
  *   <!--end::Input-->
  *   <!--begin::Button-->
  *   <button id="button" class="btn btn-icon btn-light" data-clipboard-target="#target">